+++ UniVgo2/Runtime/VgoVersion.cs
@@ -16,9 +16,9 @@ namespace UniVgo2
         public const int MINOR = 3;
 
         /// <summary>Patch</summary>
-        public const int PATCH = 0;
+        public const int PATCH = 1;
 
         /// <summary>Version</summary>
-        public const string VERSION = "2.3.0";
+        public const string VERSION = "2.3.1";
     }
 }